# docs/swagger.yaml
basePath: /
definitions:
...
  models.MenuCategory:
    properties:
      icon:
        type: string
      key:
        type: string
      label:
        type: string
      order:
        type: integer
    type: object
...
  models.PlatformService:
    properties:
      category:
        type: string
      defaultVersion:
        type: string
      description:
        type: string
      exposesUI:
        description: |-
          ExposesUI is false for infrastructure-only packages (operators, storage
          backends) that have no console page, and nil when the catalog does not set
          it, which the console treats as exposing a UI. It lets the catalog-driven
          menu leave purely infrastructural packages out of the navigation.
        type: boolean
      icon:
        type: string
      label:
        description: |-
          Label is the display name shown in the console menu; the service `name`
          (its identity and route) is used when empty.
        type: string
      name:
        type: string
      repository:
        type: string
      versions:
        items:
          type: string
        type: array
    type: object
...
  models.ProfileImage:
    properties:
      image:
        type: string
      label:
        type: string
    type: object
...
  service.ServiceVersionsResponse:
    properties:
      default:
        type: string
      versions:
        items:
          type: string
        type: array
    type: object
host: localhost:8093
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Minimal API server for OKDP UI New
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: OKDP Server New API
  version: "1.0"
paths:
...
  /api/platform-categories:
    get:
      description: Get the ordered menu sections (key, label, icon, order) the console
        groups services under
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MenuCategory'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List console navigation categories
      tags:
      - platform-services
  /api/platform-services:
    get:
      description: Get the list of managed OKDP services that can be deployed per
        project
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PlatformService'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List available platform services
      tags:
      - platform-services
    post:
      consumes:
      - application/json
      description: Expose a new managed service in the catalog (commits platform/catalog.yaml
        to the deployments repository)
      parameters:
      - description: Catalog service
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PlatformService'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PlatformService'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a service to the catalog
      tags:
      - platform-services
  /api/platform-services/{serviceName}:
    delete:
      description: Remove an exposed service from the catalog (commits platform/catalog.yaml
        to the deployments repository)
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a catalog service
      tags:
      - platform-services
    put:
      consumes:
      - application/json
      description: Update an exposed service (versions, default, description, icon,
        category)
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      - description: Catalog service
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PlatformService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlatformService'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a catalog service
      tags:
      - platform-services
  /api/platform-services/{serviceName}/inputs:
    get:
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      - description: Package version tag (defaults to the catalog default version)
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PackageInput'
            type: array
      summary: Connections a service's package declares it needs
      tags:
      - platform-services
  /api/platform-services/{serviceName}/schema:
    get:
      description: Returns the JSON Schema (parameters with UI metadata) for a given
        service version
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      - description: Package version tag (defaults to the catalog default version)
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get JSON Schema for a platform service
      tags:
      - platform-services
  /api/platform-services/{serviceName}/versions:
    get:
      description: Returns the chart versions published in the OCI registry, and the
        catalog default
      parameters:
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ServiceVersionsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List available versions for a platform service
      tags:
      - platform-services
  /api/profile-images:
    get:
      description: Returns the list of available container images per profile type
        from the platform values (global.okdp.jupyter.profiles)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.ProfileImage'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get available profile images
      tags:
      - platform-services
  /api/projects:
    get:
      consumes:
      - application/json
      description: List the projects declared in the deployments repository (projects/<p>/project.yaml),
        whether written by the console or by hand in Git
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Project'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: 'Create a project: commits projects/<name>/project.yaml and creates
        the project Namespace. 409 when the project is already declared, or when a
        Namespace of that name exists and is not a project''s'
      parameters:
      - description: Project Object
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Invalid project name
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Project or Namespace already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a project
      tags:
      - projects
  /api/projects/{name}:
    delete:
      consumes:
      - application/json
      description: 'Delete a project: removes projects/<name>/ from the deployments
        repository (the GitOps engine uninstalls its releases), then its Namespace
        if the console created it'
      parameters:
      - description: Project Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a project
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Get a project declared in the deployments repository (projects/<name>/project.yaml)
      parameters:
      - description: Project Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update a project's mutable metadata (currently its description)
        in projects/<name>/project.yaml; the file's other keys are kept
      parameters:
      - description: Project Name
        in: path
        name: name
        required: true
        type: string
      - description: Project Object
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a project
      tags:
      - projects
  /api/projects/{name}/connections:
    get:
      description: Connections declared in the project (projects/<p>/connections/<name>.yaml
        in the deployments repository). Connections published by deployed instances
        are returned by the internal endpoint.
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ConnectionResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the external connections of a project
      tags:
      - connections
    post:
      consumes:
      - application/json
      description: Stores the credential fields in a Kubernetes Secret of the project
        namespace and commits the connection file referencing it (secretRef) to the
        deployments repository
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      - description: Connection
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConnectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ConnectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an external connection in a project
      tags:
      - connections
  /api/projects/{name}/connections/{connName}:
    delete:
      description: Removes the connection file from the deployments repository and
        the Secret the console wrote for it. 400 while an instance still layers the
        connection in.
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      - description: Connection name
        in: path
        name: connName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an external connection of a project
      tags:
      - connections
    put:
      consumes:
      - application/json
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      - description: Connection name
        in: path
        name: connName
        required: true
        type: string
      - description: Connection
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConnectionResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an external connection of a project
      tags:
      - connections
  /api/projects/{name}/connections/{connName}/consumers:
    get:
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Connection name
        in: path
        name: connName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ConnectionConsumer'
            type: array
      summary: Services of the project bound to a connection
      tags:
      - connections
  /api/projects/{name}/connections/internal:
    get:
      description: Connections provided by the services already deployed in the project
        (the outputs of their instance descriptor), that the project's other services
        can consume
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.InternalConnection'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List the internal connections of a project
      tags:
      - connections
  /api/projects/{name}/connections/selectable:
    get:
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Contract the connections must satisfy
        in: query
        name: contract
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SelectableConnection'
            type: array
      summary: Connections offered for a package input of a given contract
      tags:
      - connections
  /api/projects/{name}/connections/test:
    post:
      consumes:
      - application/json
      description: 'Opens a real connection to the endpoint and checks the credentials.
        Nothing is written to the cluster, so a connection can be validated before
        being created. Always answers 200: the outcome is in the body.'
      parameters:
      - description: Project name (= Kubernetes namespace)
        in: path
        name: name
        required: true
        type: string
      - description: Connection settings to probe
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ConnectionTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ConnectionTestResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Test a connection from the platform
      tags:
      - connections
...
  /api/projects/{name}/services:
    get:
      description: Get all deployed service instances for a project
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ServiceInstance'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List deployed services for a project
      tags:
      - services
    post:
      consumes:
      - application/json
      description: 'Deploy a managed platform service into a project: commits the
        instance files to the deployments Git repository. The answer carries the commit
        (revision) and the status Pending until the GitOps engine reconciles it. Only
        the submitted parameters are written to values.yaml. 400 on invalid parameters;
        409 with code instance-exists or release-name-taken.'
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service Deploy Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ServiceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ServiceInstance'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deploy a platform service
      tags:
      - services
  /api/projects/{name}/services/{serviceName}:
    delete:
      description: 'Remove a deployed service from a project: removes its directory
        from the deployments Git repository; the GitOps engine uninstalls it.'
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a deployed service
      tags:
      - services
    get:
      description: Get a specific deployed service instance
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ServiceInstance'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a deployed service
      tags:
      - services
...
  /api/projects/{name}/services/{serviceName}/pods:
    get:
      description: Get all pods belonging to a deployed service instance
      parameters:
      - description: Project name
        in: path
        name: name
        required: true
        type: string
      - description: Service instance name
        in: path
        name: serviceName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Pod'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List pods for a service instance
      tags:
      - services
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: |-
      OIDC access token, as "Bearer <token>". Required on
      every /api route except /api/capabilities.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"